Fixed bug where emotes were shown as text
The emote id's and corresponding index ranges in the message string given in the tags portion of the irc message are not guaranteed to be in order. Code made that assumption causing some emotes not to be displayed as images but as text. Code now checks against all remaining emotes found in the tags.