SE chat widget now expands correctly
Resizing the widget in the streamelements overlay editor should now work correctly. Before, the size was fixed and had to be modified directly in the code and the streamelements settings for that widget. Now you can just drag the border around and it will fill correctly. This does not change font size or the image sizes, just the space the messages takes up.